requirements.txt
aiosqlite==0.21.0
discord.py==2.6
humanize==4.12.3